Habit Tracker components working as intended now
